+++ Sources/PopupBridge/PopupBridgeConstants.swift
@@ -3,4 +3,5 @@ import Foundation
 class PopupBridgeConstants {
 
     static let callbackURLScheme: String = "sdk.ios.popup-bridge"
+    static let host: String = "popupbridgev1"
 }